Update assets
Add barrier/ticket machine textures. Also apply related updates to build item resources, UI prefab, scene/navmesh, and several controllers/scripts (BuildController, BuildMenuController, TrainMenuController, PreviewableObject, Train, TrainService) to integrate the new buildable.

+++ Assets/Scripts/ScriptableObjectScripts/Train.cs
@@ -22,6 +22,6 @@ public class Train : ScriptableObject
     public float secondsStationary = 10;
 
     [Header("Economy")]
-    public float upfrontCost = 100f;
-    public float costPerMinute = 10f;
+    public int upfrontCost = 100;
+    public int costPerMinute = 10;
 }
